feat: Implement Weather Carousel with custom location and station selection
- Added WeatherCarousel component to MainScreen to display weather, radar, and station data in horizontal pages
- Integrated location selection dialog for setting custom ZIP code locations
- Added station selection dialog for choosing preferred weather stations
- Updated WeatherBanner to include location button
- Updated StationDataComponent to include station selection button
- Connected UI components with WeatherViewModel for data management
- Added proper state handling for dialog display

TODO
make the horizontal cards carosail min height, the size of the largest card, and make all the other cards match, having the cards of different heights causes the UX to jump around.

The Select weather station dialog box does not have any weather staions loaded in it to select. it should populate a list of the closest 5 to 10 staions.